titan_veritas/scrapers/youth_leagues/michigan_sysa.py
"""Scaffold scraper for Michigan State Youth Soccer Association (MSYSA).

Target: https://www.michiganyouthsoccer.org
Focus: Detroit/Troy area clubs — high-density San Marino diaspora hub.

STATUS: Scaffold — URL patterns and selectors defined, real implementation
requires manual verification of site structure.
"""

import logging
from typing import List, Optional

import httpx
from bs4 import BeautifulSoup

from ...core.models import PlayerProfile
from ...core.rate_limiter import RateLimiter
from .base import AsyncLeagueScraper

logger = logging.getLogger(__name__)


class MichiganSYSAScraper(AsyncLeagueScraper):
    """Scraper for Michigan State Youth Soccer Association rosters."""

    @property
    def league_name(self) -> str:
        return "Michigan SYSA"

    @property
    def country(self) -> str:
        return "USA"

    @property
    def base_url(self) -> str:
        return "https://www.michiganyouthsoccer.org"

    # Focus cities in the Detroit metro area (SM diaspora cluster)
    TARGET_CITIES = ["Detroit", "Troy", "Sterling Heights", "Warren", "Dearborn"]

    async def search_by_surname(self, surname: str) -> List[PlayerProfile]:
        """Search MSYSA player registry by surname.

        NOTE: This is a scaffold. The actual MSYSA site may require
        authentication or have a different search mechanism.
        Raises NotImplementedError until the real site structure is verified.
        """
        profiles = []

        async with httpx.AsyncClient(timeout=30.0, follow_redirects=True) as client:
            # Attempt roster search
            search_url = f"{self.base_url}/rosters/search"
            response = await self._fetch(client, search_url, params={"q": surname})

            if response is None:
                logger.info(f"[Michigan SYSA] Could not access roster search for {surname}")
                return profiles

            soup = BeautifulSoup(response.text, "html.parser")
            rows = soup.select("table.roster tr, .player-card, .roster-entry")

            for row in rows:
                try:
                    name_el = row.select_one(".player-name, td:first-child")
                    if not name_el:
                        continue

                    full_name = name_el.get_text(strip=True)
                    parts = full_name.split()
                    if len(parts) < 2:
                        continue

                    # Check surname match
                    last_name = parts[-1]
                    if surname.lower() not in last_name.lower():
                        continue

                    age_el = row.select_one(".age, td.age")
                    age = None
                    if age_el:
                        try:
                            age = int(age_el.get_text(strip=True))
                        except ValueError:
                            pass

                    club_el = row.select_one(".club, td.team")
                    club = club_el.get_text(strip=True) if club_el else "Michigan Youth"

                    profiles.append(PlayerProfile(
                        first_name=parts[0],
                        last_name=last_name,
                        known_as=full_name,
                        age=age,
                        birth_country="USA",
                        birth_city="Detroit Metro",
                        nationalities=["USA"],
                        current_club=club,
                        current_league="Michigan SYSA",
                        source="Youth League (Michigan SYSA)",
                        source_url=self.base_url,
                    ))

                except Exception as e:
                    logger.debug(f"[Michigan SYSA] Parse error: {e}")
                    continue

        return profiles

    async def search_by_age_range(self, min_age: int, max_age: int) -> List[PlayerProfile]:
        """Search by age range — not available for this league without full roster access."""
        raise NotImplementedError(
            "Michigan SYSA age-range search requires authenticated roster access"
        )
