requirements.txt
httpx
beautifulsoup4
jinja2
click
rich
tqdm
# Phonetic & fuzzy matching (used by RosettaStone)
jellyfish
thefuzz[speedup]
python-Levenshtein
# Environment
python-dotenv
# Gmail API
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
# Retry logic with backoff
tenacity
